Move fixtureTree functions from playground2 package to shared2 package
