+++ packages/react-cosmos-playground2/src/plugins/Root/RendererHeader.tsx
@@ -1,5 +1,6 @@
 import { isEqual } from 'lodash';
 import React from 'react';
+import { TreeNode } from 'react-cosmos-shared2/fixtureTree';
 import { FixtureId } from 'react-cosmos-shared2/renderer';
 import styled from 'styled-components';
 import {
@@ -9,7 +10,6 @@ import {
   XCircleIcon
 } from '../../shared/icons';
 import { RendererActionSlot } from '../../shared/slots/RendererActionSlot';
-import { TreeNode } from '../../shared/tree';
 import { IconButton32 } from '../../shared/ui/buttons';
 import { grey176, grey32, white10 } from '../../shared/ui/colors';
