--- packages/react-cosmos-playground2/src/shared/fixtureTree/shared.ts
@@ -1,5 +1,0 @@
-import { FixtureId } from 'react-cosmos-shared2/renderer';
-import { TreeNode, TreeNodeDirs } from '../tree';
-
-export type FixtureNode = TreeNode<FixtureId>;
-export type FixtureNodeDirs = TreeNodeDirs<FixtureId>;
