+++ packages/react-cosmos-playground2/src/shared/ui/valueInputTree/shared.ts
@@ -3,9 +3,9 @@ import {
   FixtureStatePrimitiveValue,
   FixtureStateUnserializableValue
 } from 'react-cosmos-shared2/fixtureState';
+import { TreeNode } from 'react-cosmos-shared2/fixtureTree';
 import { FixtureId } from 'react-cosmos-shared2/renderer';
 import styled from 'styled-components';
-import { TreeNode } from '../../tree';
 
 export type TreeItemValue =
   | FixtureStatePrimitiveValue
