+++ packages/react-cosmos-shared2/fixtureTree.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/fixtureTree';
