+++ packages/react-cosmos-shared2/fixtureTree.ts
@@ -1,0 +1,1 @@
+export * from './dist/fixtureTree';
