+++ packages/react-cosmos-shared2/src/fixtureTree/collapseDirs.test.ts
