+++ packages/react-cosmos-shared2/src/fixtureTree/collapseDirs.ts
@@ -1,5 +1,5 @@
 import { forEach } from 'lodash';
-import { FixtureNode, FixtureNodeDirs } from './shared';
+import { FixtureNode, FixtureNodeDirs } from './shared/types';
 
 export function collapseDirs(
   treeNode: FixtureNode,
