+++ packages/react-cosmos-shared2/src/fixtureTree/collapseSoloIndexes.test.ts
