+++ packages/react-cosmos-shared2/src/fixtureTree/collapseSoloNamedItems.test.ts
