+++ packages/react-cosmos-shared2/src/fixtureTree/createRawFixtureTree.test.ts
