+++ packages/react-cosmos-shared2/src/fixtureTree/createRawFixtureTree.ts
@@ -1,6 +1,6 @@
 import { get, set } from 'lodash';
-import { FixtureId, FixtureNamesByPath } from 'react-cosmos-shared2/renderer';
-import { FixtureNode } from './shared';
+import { FixtureId, FixtureNamesByPath } from '../renderer';
+import { FixtureNode } from './shared/types';
 
 export function createRawFixtureTree(
   fixtures: FixtureNamesByPath
