+++ packages/react-cosmos-shared2/src/fixtureTree/hideFixtureSuffix.test.ts
