+++ packages/react-cosmos-shared2/src/fixtureTree/hideFixtureSuffix.ts
@@ -1,5 +1,5 @@
 import { mapKeys } from 'lodash';
-import { FixtureNode } from './shared';
+import { FixtureNode } from './shared/types';
 
 export function hideFixtureSuffix(
   treeNode: FixtureNode,
