+++ packages/react-cosmos-shared2/src/fixtureTree/hideSingleChildDirs.test.ts
