+++ packages/react-cosmos-shared2/src/fixtureTree/shared/types.ts
@@ -1,0 +1,13 @@
+import { FixtureId } from '../../renderer';
+
+export type TreeNode<Item> = {
+  items: { [itemName: string]: Item };
+  dirs: TreeNodeDirs<Item>;
+};
+
+export type TreeNodeDirs<Item> = {
+  [dirName: string]: TreeNode<Item>;
+};
+
+export type FixtureNode = TreeNode<FixtureId>;
+export type FixtureNodeDirs = TreeNodeDirs<FixtureId>;
